# Implement EventList generateHistogram without sorting, use within AlignAndFocusPowder