--- a/Framework/DataObjects/inc/MantidDataObjects/EventList.h
+++ b/Framework/DataObjects/inc/MantidDataObjects/EventList.h
@@ -205,8 +205,8 @@ public:
                          EventList *destination);
   // get EventType declaration
   void generateHistogram(const MantidVec &X, MantidVec &Y, MantidVec &E, bool skipError = false) const override;
-  void generateHistogram(const double xmin, const double step, const double xmax, const MantidVec &X, MantidVec &Y,
-                         MantidVec &E, bool skipError = false) const;
+  void generateHistogram(const double step, const MantidVec &X, MantidVec &Y, MantidVec &E,
+                         bool skipError = false) const;
   void generateHistogramPulseTime(const MantidVec &X, MantidVec &Y, MantidVec &E,
                                   bool skipError = false) const override;
 
@@ -354,8 +354,7 @@ private:
                                                                      const double &tofOffset) const;
 
   void generateCountsHistogram(const MantidVec &X, MantidVec &Y) const;
-  void generateCountsHistogram(const double xmin, const double step, const double xmax, const MantidVec &X,
-                               MantidVec &Y) const;
+  void generateCountsHistogram(const double step, const MantidVec &X, MantidVec &Y) const;
 
   static size_t findLinearBin(const MantidVec &X, const double xmin, const double step, const double tof);
   static size_t findLogBin(const MantidVec &X, const double tof, const double divisor, const double offset);
@@ -387,8 +386,8 @@ private:
   template <class T>
   static void histogramForWeightsHelper(const std::vector<T> &events, const MantidVec &X, MantidVec &Y, MantidVec &E);
   template <class T>
-  static void histogramForWeightsHelper(const std::vector<T> &events, const double xmin, const double step,
-                                        const double xmax, const MantidVec &X, MantidVec &Y, MantidVec &E);
+  static void histogramForWeightsHelper(const std::vector<T> &events, const double step, const MantidVec &X,
+                                        MantidVec &Y, MantidVec &E);
   template <class T>
   static void integrateHelper(std::vector<T> &events, const double minX, const double maxX, const bool entireRange,
                               double &sum, double &error);