--- a/docs/source/release/v6.9.0/Diffraction/Powder/New_features/36136.rst
+++ b/docs/source/release/v6.9.0/Diffraction/Powder/New_features/36136.rst
@@ -1,0 +1,1 @@
+- :ref:`AlignAndFocusPowder <algm-AlignAndFocusPowder>` will no longer sort events prior to doing :ref:`RebinRagged <algm-RebinRagged>` when ``PreserveEvents=False`` enabling it to ultilize performace improvement to :ref:`Rebin <algm-Rebin>`.