--- a/docs/source/release/v6.9.0/Framework/Algorithms/New_features/36136.rst
+++ b/docs/source/release/v6.9.0/Framework/Algorithms/New_features/36136.rst
@@ -1,0 +1,1 @@
+- A new method was added the histogram Event data without sorting the events first, which the :ref:`Rebin <algm-Rebin>` algorithm now uses when ``PreserveEvents=False``, giving a performance improvement.