… Framework/Algorithms/test/RebinTest.h
… Framework/DataObjects/test/EventListTest.h
… Framework/WorkflowAlgorithms/src/AlignAndFocusPowder.cpp
… buildconfig/CMake/CppCheck_Suppressions.txt.in